move key value formatting to format
